删除Cache和SDK模块.
