+++ blog-api/src/main/resources/dispatcher-servlet.xml
@@ -20,4 +20,8 @@
 
     <bean class="com.zuoxiaolong.blog.common.spring.JsonHandlerExceptionResolver"/>
 
+    <bean class="org.springframework.web.multipart.commons.CommonsMultipartResolver">
+        <property name="maxUploadSize" value="5400000"/>
+    </bean>
+
 </beans>
