--- blog-cache/src/main/java/com/zuoxiaolong/blog/cache/constants/CacheConstants.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2002-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.zuoxiaolong.blog.cache.constants;
-
-/**
- * 缓存模块常量类
- *
- * @author goozi
- * @create 2016-05-17 19:57
- * @since 1.0.0
- */
-public class CacheConstants {
-    //推荐
-    public static final String ACTION_TYPE_RECOMMEND = "1";
-    //阅读
-    public static final String ACTION_TYPE_READ = "2";
-    //评论
-    public static final String ACTION_TYPE_COMMEND = "3";
-}
