+++ blog-fast/src/main/webapp/WEB-INF/web.xml
@@ -5,4 +5,8 @@
 		  http://java.sun.com/xml/ns/javaee/web-app_3_0.xsd"
          version="3.0">
 
+    <welcome-file-list>
+        <welcome-file>/index.html</welcome-file>
+    </welcome-file-list>
+
 </web-app>
