+++ blog-fast/src/main/webapp/bootstrap/js/bootstrap.js
