+++ blog-fast/src/main/webapp/bootstrap/js/npm.js
