--- blog-sdk/pom.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>everyone-java-blog</artifactId>
-        <groupId>com.zuoxiaolong.blog</groupId>
-        <version>1.0.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>blog-sdk</artifactId>
-    <name>${project.groupId}:${project.artifactId}</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.zuoxiaolong.blog</groupId>
-            <artifactId>blog-model</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.zuoxiaolong.blog</groupId>
-            <artifactId>blog-common</artifactId>
-        </dependency>
-    </dependencies>
-</project>
