--- blog-sdk/src/main/java/com/zuoxiaolong/blog/sdk/exception/BlogSdkPropertiesUtilException.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2002-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.zuoxiaolong.blog.sdk.exception;
-
-/**
- * @author Boren You
- * @date 2016/5/26 18:32
- * @since 1.0.0
- */
-public class BlogSdkPropertiesUtilException extends Exception {
-
-    public BlogSdkPropertiesUtilException(String message) {
-        super(message);
-    }
-
-    public BlogSdkPropertiesUtilException(Exception cause) {
-        super(cause);
-    }
-
-    public BlogSdkPropertiesUtilException(String message, Exception cause) {
-        super(message, cause);
-    }
-}
