--- blog-sdk/src/test/java/com/zuoxiaolong/blog/sdk/BlogInterfaceHandlerTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2002-2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.zuoxiaolong.blog.sdk;
-
-import com.zuoxiaolong.blog.model.dto.cache.ArticleRankResponseDto;
-
-import java.util.List;
-
-/**
- * @author Boren You
- * @dateTime 2016/5/26 22:18
- * @since 1.0.0
- */
-public class BlogInterfaceHandlerTest {
-    public static void main(String[] args) throws Exception {
-        List<ArticleRankResponseDto> list = BlogSdk.getArticleRank();
-    }
-}
