--- blog-sdk/src/test/resources/blog-sdk.properties
@@ -1,1 +1,0 @@
-articleRankUrl=http://localhost:8083/cache/article/rank
