+++ pom.xml
@@ -37,8 +37,6 @@
         <module>blog-admin</module>
         <module>blog-common</module>
         <module>blog-model</module>
-        <module>blog-cache</module>
-        <module>blog-sdk</module>
         <module>blog-fast</module>
     </modules>
